Move check_cudnn() and jurigged setup to startup_utils.py.
