+++ invokeai/app/util/startup_utils.py
@@ -1,4 +1,9 @@
+import logging
 import socket
+
+import torch
+
+from invokeai.backend.util.logging import InvokeAILogger
 
 
 def find_open_port(port: int) -> int:
@@ -11,3 +16,31 @@ def find_open_port(port: int) -> int:
             return find_open_port(port=port + 1)
         else:
             return port
+
+
+def check_cudnn(logger: logging.Logger) -> None:
+    """Check for cuDNN issues that could be causing degraded performance."""
+    if torch.backends.cudnn.is_available():
+        try:
+            # Note: At the time of writing (torch 2.2.1), torch.backends.cudnn.version() only raises an error the first
+            # time it is called. Subsequent calls will return the version number without complaining about a mismatch.
+            cudnn_version = torch.backends.cudnn.version()
+            logger.info(f"cuDNN version: {cudnn_version}")
+        except RuntimeError as e:
+            logger.warning(
+                "Encountered a cuDNN version issue. This may result in degraded performance. This issue is usually "
+                "caused by an incompatible cuDNN version installed in your python environment, or on the host "
+                f"system. Full error message:\n{e}"
+            )
+
+
+def enable_dev_reload() -> None:
+    """Enable hot reloading on python file changes during development."""
+    try:
+        import jurigged
+    except ImportError as e:
+        raise RuntimeError(
+            'Can\'t start `--dev_reload` because jurigged is not found; `pip install -e ".[dev]"` to include development dependencies.'
+        ) from e
+    else:
+        jurigged.watch(logger=InvokeAILogger.get_logger(name="jurigged").info)
